+++ packages/css-calc/src/util/patch-result.ts
@@ -9,17 +9,17 @@ import type { conversionOptions } from '../options';
 export function patchCalcResult(x: TokenNode | FunctionNode | -1, options?: conversionOptions): TokenNode | FunctionNode | -1 {
 	let y: TokenNode | FunctionNode | -1 = x;
 
-	if (!options?.censorIntoStandardRepresentableValues) {
-		y = patchNaN(y);
-		y = patchInfinity(y);
-	}
-
 	if (options?.toCanonicalUnits) {
 		y = patchCanonicalUnit(y);
 	}
 
 	y = patchPrecision(y, options?.precision);
 	y = patchMinusZero(y);
+
+	if (!options?.censorIntoStandardRepresentableValues) {
+		y = patchNaN(y);
+		y = patchInfinity(y);
+	}
 
 	return y;
 }